refactor: rename card directory name
