+++ apps/web/src/modules/job/components/Card/components/Footer/styled.ts
