+++ apps/web/src/modules/job/components/Card/components/Footer/type.ts
