+++ apps/web/src/modules/job/components/Card/components/Header/type.ts
