+++ apps/web/src/modules/job/components/Card/components/ShootingDetail/index.tsx
