+++ apps/web/src/modules/job/components/Card/components/ShootingDetail/styled.ts
