+++ apps/web/src/modules/job/components/Card/styled.ts
