+++ apps/web/src/modules/job/components/Card/type.ts
